% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_participants}
\alias{ra_participants}
\title{Ribosome accumulation-study: Participants}
\format{
A data frame with 19 rows and 8 variables:
\describe{
\item{participant}{Participant id}
\item{group}{experiment, Experimental group; control, non-training control group}
\item{age}{age in years}
\item{height}{body stature in cm}
\item{weight}{body mass in kg}
\item{sex}{M, male; F, female}
\item{R}{Training condition allocated to right leg: const, constant volume
(6 sets of knee-extension) over the whole training period; var, variable
volume ov the training period (6 sets during session 1-4, 3 sets during
sessions 5-8 and 9 sets during sessions 9-12.); ctrl_leg, control condition leg.}
\item{L}{Training condition allocated to left leg: const, constant volume
(6 sets of knee-extension) over the whole training period; var, variable
volume ov the training period (6 sets during session 1-4, 3 sets during
sessions 5-8 and 9 sets during sessions 9-12.); ctrl_leg, control condition leg.}
}
}
\usage{
ra_participants
}
\description{
Participant sex, baseline characteristics, group and leg allocation.
}
\keyword{datasets}
