% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_qpcr}
\alias{ra_qpcr}
\title{Ribosome accumulation study: Quantitative PCR data}
\format{
A data frame with 9613 rows and 10 variables:
\describe{
\item{participant}{Participant identification}
\item{leg}{R, right; L, left}
\item{time}{Sampling time-point. S0, before the intervention; S1 S4, S5,
S8, S9 and S12, 48 hours after sessions 1, 4, 5, 8, 9, and 12; post1w,
8 days after session 12; S1c 48 h control period; postctrl, 2-4 week control
period}
\item{sex}{Participant sex. Male, M; Female, F.}
\item{cond}{Experimental condition}
\item{target}{Gene target (see ra_reagents for details)}
\item{cq}{Quantification cycle, corresponding to the second derivate
maximum (see Spiess for details)}
\item{eff}{Amplification efficiency per target, target average.}
\item{cdna}{cDNA preparation (1 or 2)}
\item{tissue_weight}{Weight (mg) of tissue used for analysis}
}
}
\usage{
ra_qpcr
}
\description{
Relative cDNA abundance of selected RNA-targets.
}
\keyword{datasets}
