% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_tissuesamples}
\alias{ra_tissuesamples}
\title{Tissue samples for RNA extraction}
\format{
A data frame with 269 rows and 6 variables:
\describe{
\item{participant}{participant identification}
\item{leg}{R, right; L, left}
\item{time}{Sampling time-point. S0, before the intervention; S1 S4, S5,
S8, S9 and S12, 48 hours after sessions 1, 4, 5, 8, 9, and 12; post1w,
8 days after session 12; S1c 48 h control period; postctrl, 2-4 week control
period}
\item{series}{Extraction series. A subset of participants had replicate tissue
samples extracted these are represented as 1 and 2 udner series}
\item{sample}{Sample identification number. Samples were randomly assigned
an id during RNA extraction and processing.}
\item{cond}{Experimental condition}
}
}
\usage{
ra_tissuesamples
}
\description{
Tissue samples for RNA extraction
}
\keyword{datasets}
