% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_totalrna}
\alias{ra_totalrna}
\title{Total RNA and muscle weight data}
\format{
A data frame with 1134 rows and 7 variables:
\describe{
\item{participant}{participant identification}
\item{series}{Extraction series. A subset of participants had replicate tissue
samples extracted these are represented as 1 and 2 udner series}
\item{sample}{Sample identification number. Samples were randomly assigned
an id during RNA extraction and processing.}
\item{leg}{R, right; L, left}
\item{well}{Well for spectrophotometrical evaluation of RNA concentration}
\item{RNA}{Estimated total amount of RNA extracted from muscle samples (ng)}
\item{weight}{Muscle tissue weight (mg)}
}
}
\usage{
ra_totalrna
}
\description{
Total RNA and muscle weight data
}
\keyword{datasets}
